\name{Bahrami1} 
\docType{data} 
\alias{Bahrami1}
\title{Body Mass Index Data} 
\description{The medical data set is obtained from an observational study in
 the Taleghani  hospital in Tehran. These data record the 
Steatosis and BMI for 61 diabetic patients.  BMI is a statistical measure of the
weight of body mass index. A person scaled height body mass index
may be accurately calculated using any of the formulas such as
 BMI=w/H^2 where W is weight and H is height. As
explanatory variable for BMI, duration of diabet
is an amount of time or a particular time interval which a person
take diabet (a metabolic disorder characterized by high blood suger and other signs). another
 explanatory variable is gender.} 
\usage{data(Bahrami1)} 
\format{A data frame with 15 observations on the following 3 variables:DURATION,BMI,GENDER} 
\source{The medical data set is obtained from an observational study in
 the Taleghani  hospital in Tehran.} 
\references{Bahrami Samani, E.,  Ganjali, M. and Khodaddadi, A.  (2008). A Latent Variable Model for Mixed Continuous and Ordinal Responses. Journal of Statistical Theory and Applications. 7(3):337-349. } 
\keyword{Regression}